+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog of threedigrid-builder
 0.11.0 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Accept dist_calc_points <= 0; the effect is that there are no interpolated nodes.
 
 
 0.10.0 (2021-12-21)